requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.5.0
asgiref==3.8.1
bidict==0.23.1
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.8.2
configobj==5.0.9
configparser==7.1.0
distro==1.9.0
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
etelemetry==0.3.1
fastapi==0.115.0
filelock==3.16.1
fitz==0.0.1.dev2
flower==2.0.1
fsspec==2024.9.0
gevent==24.2.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.25.0
humanize==4.10.0
idna==3.10
isodate==0.6.1
jiter==0.5.0
kombu==5.4.2
looseversion==1.3.0
lxml==5.3.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==2.1.1
openai==1.46.1
packaging==24.1
pandas==2.2.3
pathlib==1.0.1
prometheus_client==0.21.0
prompt_toolkit==3.0.47
prov==2.0.1
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
redis==5.0.8
requests==2.32.3
scipy==1.14.1
serpy==0.3.1
setuptools==75.1.0
simple-websocket==1.0.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.5
tokenizers==0.20.0
tornado==6.4.1
tqdm==4.66.5
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.3
